requirements-test.txt
# a testing only requirements file for use with travis
-r requirements-full.txt
nose==1.3.4
Flask-Testing==0.4.2
